+++ src/integrations/tencent-meeting/index.ts
@@ -1,5 +1,16 @@
+/*
+ * @Author: 杨仕明 shiming.y@qq.com
+ * @Date: 2025-12-28 11:37:14
+ * @LastEditors: 杨仕明 shiming.y@qq.com
+ * @LastEditTime: 2026-01-03 08:39:30
+ * @FilePath: /lulab_backend/src/integrations/tencent-meeting/index.ts
+ * @Description: 
+ * 
+ * Copyright (c) 2026 by LuLab-Team, All Rights Reserved. 
+ */
+
 export * from './tencent.module';
 export * from './api.service';
 export * from './crypto.util';
-export * from './types';
+export * from './types/index';
 export * from './exceptions';
