+++ src/integrations/tencent-meeting/types/index.ts
@@ -1,0 +1,6 @@
+export * from './recording.types';
+export * from './meeting-record.types';
+export * from './meeting-detail.types';
+export * from './participant.types';
+export * from './smart-summary.types';
+export * from './transcript.types';
